backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.0
async-timeout==5.0.1
attrs==25.4.0
blinker==1.9.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.1.8
distro==1.9.0
exceptiongroup==1.3.1
Flask==3.0.0
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
frozenlist==1.8.0
gunicorn==21.2.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.12.0
MarkupSafe==3.0.3
multidict==6.7.0
numpy==2.0.2
openai==1.51.0
packaging==25.0
propcache==0.4.1
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.8.0
python-dotenv==1.0.0
requests==2.32.5
sniffio==1.3.1
SQLAlchemy==2.0.44
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.4
yarl==1.22.0
zipp==3.23.0
